+++ src/components/containers/index.js
@@ -4,4 +4,4 @@ export { default as AllCampusesContainer } from "./AllCampusesContainer";
 export { default as CampusContainer } from "./CampusContainer";
 export { default as AllStudentsContainer } from "./AllStudentsContainer";
 export { default as StudentContainer } from "./StudentContainer";
-export { default as NewCampusContainer } from './NewCampusContainer';
+export { default as AddCampusContainer } from './AddCampusContainer';
